pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vulnhunter"
version = "1.0.0"
description = "Autonomous AI-powered penetration testing platform"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "VulnHunter Team"}]
keywords = ["security", "penetration-testing", "ai", "vulnerability-scanner"]
dependencies = [
    "openai>=1.0.0",
    "aiohttp>=3.9.0",
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.28",
    "websockets>=12.0",
    "python-whois>=0.9",
    "dnspython>=2.6",
    # Phase 1: Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    # Phase 2: Auth
    "bcrypt>=4.0",
    "python-jose[cryptography]>=3.3",
    # Phase 4: Container isolation
    "aiodocker>=0.22",
    # Phase 5: Memory
    "pgvector>=0.3",
    # Phase 6: OSINT
    "beautifulsoup4>=4.12",
    # Phase 7: Reporting
    "reportlab>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
]

[project.scripts]
vulnhunter = "vulnhunter.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
